pyproject.toml
[tool.poetry]
name = "Travel Agency"
version = "0.1.0"
description = "AI Travel Planner"
authors = ["Arup Sarkar, Rahul Kumar <arup.sarkar@globallogic.com>, <rahul.kumar31@globallogic.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"
crewai = "^0.11.0"
duckduckgo-search = "^4.4.2"
python-dotenv = "^1.0.1"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.3"

unstructured = '==0.10.25'
pyowm = '3.3.0'
langchain-openai = "^0.0.5"
boto3 = "^1.34.64"
pydantic = "^2.6.4"
langchain-community = "^0.0.29"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"